bug 1483246: extract exploitability report into separate app
