+++ webapp-django/crashstats/crashstats/urls.py
@@ -27,9 +27,6 @@ urlpatterns = [
     url(r'^crontabber-state/$',
         views.crontabber_state,
         name='crontabber_state'),
-    url(r'^exploitability/$',
-        views.exploitability_report,
-        name='exploitability_report'),
     url(r'^report/index/(?P<crash_id>[\w-]+)$',
         views.report_index,
         name='report_index'),
