+++ webapp-django/crashstats/exploitability/urls.py
@@ -1,0 +1,10 @@
+from django.conf.urls import url
+from crashstats.exploitability import views
+
+
+app_name = 'exploitability'
+urlpatterns = [
+    url(r'^exploitability/$',
+        views.exploitability_report,
+        name='report'),
+]
